feat: add AverageChart data
