+++ src/components/AverageChartWrap.jsx
@@ -3,7 +3,6 @@ import styled from "styled-components";
 
 import CardTemplate from "./CardTemplate";
 import AverageChart from "./AverageChart";
-import Chart from "./Chart";
 
 const AverageChartWrap = () => {
   // 평균점수 Wrap
@@ -21,7 +20,7 @@ const AverageChartWrap = () => {
     <Cont>
       {/* 카드 테두리 Components */}
       <CardTemplate
-        Element={Chart}
+        Element={AverageChart}
         Name={"평균점수"}
         Count={"+ 2.1%"}
         Info={" 전달 대비 평균 총점"}
